Add non-hacked Oracle type mappings for bool, DateTimeOffset, and Guid
Issue #9711 and part of #242

Guild and bool required just type conversions (and some fixes in the product code).

DateTimeOffset cannot be read for the store without using an Oracle-specific call on the DbDataReader service. To enable this there is now an IDataReaderMethodProvider service. This can then be used with a type conversion to get the correct behavior.

There are still some hacks remaining for other areas--specifically, the cursor parameter hack and some SQL manipulation.